Scripts: Make main run script more robust
The run would fail if IAPS was not loaded, even when the experiment didn't
need IAPS. This was fixed. Now the experiment script can be run, even
when the custom module is not loaded.

+++ run_mini_project.sh
@@ -160,8 +160,10 @@ then
 	type="IAPS"
 else
 	echo "Disable Custom"
-	echo 0 > /sys/module/pkt_steer_module/parameters/custom_toggle
-
+	if test -f /sys/module/pkt_steer_module/parameters/custom_toggle
+	then 
+		echo 0 > /sys/module/pkt_steer_module/parameters/custom_toggle
+	fi
 	#rmmod pkt_steer_module
 	#if [[ "$rfs" == "0" ]]
 	#then
